pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["AM <alexandre.meline.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
markdown = "^3.7"
django-filter = "^24.3"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"
psycopg2 = "^2.9.9"
django-redis = "^5.4.0"
django-split-settings = "^1.3.2"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
